intified damage

+++ damage.h
@@ -4,7 +4,7 @@
 #include "attr.h"
 #include "noob.h"
 
-void damage_calc(noob_t *noob, float damage, float dt, attr_t attr);
+void damage_calc(noob_t *noob, int damage, int dt, attr_t attr);
 int damage_not_worthwhile(noob_t *noob, attr_t attr);
 
 #endif
